Add files via upload
Easton, Washington is a small town located next to Interstellar 90 in WA that many people have not heard of. The K-12 school and playground is separated from the freeway by a chain link fence. At the playground the decibel level 94 decibels, and the safe limit 85 decibels which means we are at risk of temporary or permanent hearing loss. For a solution we want to build a wall. A wall can be used as a sound barrier. According to the Washington State Department of Transportation some strategies for blocking out sounds is a barrier, “some alternatives departed from more conventional use of noise barriers, and include sound absorption applied to other design features.” If we were to build a wall we could block out sound and decrease the noise decibel so we wouldn't be at risk.